% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{breakdown}
\alias{breakdown}
\title{breakdown}
\usage{
breakdown(x)
}
\arguments{
\item{x}{is a character string}
}
\value{
the reduced character string
}
\description{
breakdown function from LSAfun 0.6.2 to remove special characters from a character string
}
