% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{gg_color_hue}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{number of colors needed}
}
\value{
returns a vector of colors for use in plotting
}
\description{
internal function that acts as a ggplot color replicator helper function
}
\examples{
\dontrun{
gg_color_hue(8)
}
}
