% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_visualization.R
\name{GO_viz_choose}
\alias{GO_viz_choose}
\title{This function provides a dot plot chart of the GO enrichments of user defined Parent term categories from the results of GO_visualizatiohn function}
\usage{
GO_viz_choose(
  GO_viz_results,
  chosen_cats,
  clust_list = NULL,
  markers_df = NULL,
  GOcats = NULL,
  species_x = "Homo sapiens",
  goterms = NULL
)
}
\arguments{
\item{GO_viz_results}{object provided by the Go_visulaization function}

\item{chosen_cats}{User supplied catergorical terms to plot on x-axis. Default is null which lets the function choose the categories.}

\item{clust_list}{a list of gene vectors associated with each cluster.}

\item{markers_df}{data frame resulting from FindAllMarkers function in the Seurat package, required if a clust_list is not provided}

\item{GOcats}{product of the msigdb_gsets funciton using GOBP. Can also be user supplied annotation in the same format.}

\item{species_x}{species names, default is "Homo sapiens", refer to ?msigdbr::msigdbr for available species}

\item{goterms}{a vector with GOIDS as object and the GO terms as names}
}
\value{
prints a plot and returns a ggplot2 object suitable for fine tuning the appearance of the plot.
}
\description{
This function provides a dot plot chart of the GO enrichments of user defined Parent term categories from the results of GO_visualizatiohn function
}
\examples{
\dontrun{
GO_viz_choose(go_vis_res1,chosen_cats=newcats,markers_df=markers)
}
}
