% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{markers}
\alias{markers}
\title{markers}
\format{
a data frame that is the result of the FIndAllMarkers function in Seurat
}
\usage{
markers
}
\description{
A dataset from 10X genomics containing single cell RNA-seq data from ~3K peripheral blood mononuclear cells  analyzed for cluster markers.
}
\keyword{datasets}
