% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{prepGOTERMS}
\alias{prepGOTERMS}
\title{prepGOTERMS}
\usage{
prepGOTERMS(x)
}
\arguments{
\item{x}{is the vector from Term(GOTERM) using the GO.db package}
}
\description{
function to generate a GO vector from the GO.db
}
