% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{reactome_prep}
\alias{reactome_prep}
\title{reactome_prep}
\usage{
reactome_prep(cluster_enriched_df, RP, RP_adj)
}
\arguments{
\item{cluster_enriched_df}{data frame produced by enrich_test function}

\item{RP}{product of RPprep function}

\item{RP_adj}{Pathway relationship file from Reactome. To decrease time required should limit to organism of interest
by selecting for organism, e.g. RPH<-RP[grep("HSA",RPR[,1]),]. Can be modified by user
to use some 2nd level terms (e.g. RP_adj<-RPH[-(grep("R-HSA-8953897",RPH[,1])),]  ).}
}
\value{
returns a data frame for use in the reactome_visualization function
}
\description{
function to make a data to use with visualization function
}
\examples{
RP_ready<-reactome_prep(enrich_results$Enriched_df,RP=RP,RP_adj=mRPR)
}
