% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_visualization.R
\name{reactome_visualization}
\alias{reactome_visualization}
\title{reactome_visualization}
\usage{
reactome_visualization(RP.df, RP_adj, RP = NULL, path_order = NULL)
}
\arguments{
\item{RP.df}{data frame produced by reactome_prep function}

\item{RP_adj}{same data frame as used in reactome_prep function}

\item{RP}{data frame of Path IDS from Reactome database. IF left null uses the RP file included as data in the screp package.}

\item{path_order}{a vector of parent pathway names from the rootName column in the RP.df. Can be used to
plot the parent pathway names in desired order along the x-axis. Default is NULL which gives a random order.}
}
\value{
returns a list with the first item in the list being a data frame used in creating a
         plotting object and the second item being the ggplot2 object generated
}
\description{
function to generate plot using the dataframe from the reactome_prep function
}
\examples{
\dontrun{
react_vis2<-reactome_visualization(RP_ready,mRPR,path_order=z)
}
}
