% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RP}
\alias{RP}
\title{RP}
\format{
a table read in by read.delim()
}
\usage{
RP
}
\description{
The "Complete List of Pathways" file downloaded from https://reactome.org/download-data in August 2020.
}
\keyword{datasets}
