% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RPR}
\alias{RPR}
\title{RPR}
\format{
a table read in by read.delim()
}
\usage{
RPR
}
\description{
The "Pathways hierarchy relationship" file downloaded from https://reactome.org/download-data in August 2020.
}
\keyword{datasets}
